refactor: ClassController에 있던 tutor class 관련 메서드 TutorController로 이동 및 기존 컨트롤러 삭제
