+++ src/main/java/com/linked/classbridge/service/ReviewService.java
@@ -32,11 +32,10 @@ public class ReviewService {
 
     private final ReviewImageRepository reviewImageRepository;
 
-    private final OneDayClassService oneDayClassService;
-
     private final LessonService lessonService;
 
     private final S3Service s3Service;
+    private final ClassService classService;
 
     /**
      * 리뷰 등록
@@ -217,7 +216,7 @@ public class ReviewService {
      */
     public Page<GetReviewResponse> getClassReviews(Long classId, Pageable pageable) {
 
-        OneDayClass oneDayClass = oneDayClassService.findClassById(classId);
+        OneDayClass oneDayClass = classService.findClassById(classId);
 
         Page<Review> reviews = reviewRepository.findByOneDayClass(oneDayClass, pageable);
